+++ py-polars/src/object.rs
@@ -1,81 +1,1 @@
-use std::any::Any;
-use std::borrow::Cow;
-use std::sync::Arc;
-
-use polars::prelude::*;
-use polars_core::chunked_array::object::builder::ObjectChunkedBuilder;
-use polars_core::chunked_array::object::registry;
-use polars_core::chunked_array::object::registry::AnonymousObjectBuilder;
-use polars_core::error::PolarsError::ComputeError;
-use polars_core::error::PolarsResult;
-use polars_core::frame::DataFrame;
-use pyo3::prelude::*;
-
-use crate::apply::lazy::{call_lambda_with_series, ToSeries};
-use crate::dataframe::PyDataFrame;
-use crate::prelude::{python_udf, ObjectValue};
-use crate::py_modules::POLARS;
-use crate::Wrap;
-
 pub(crate) const OBJECT_NAME: &str = "object";
-
-fn python_function_caller_series(s: Series, lambda: &PyObject) -> PolarsResult<Series> {
-    Python::with_gil(|py| {
-        let object = call_lambda_with_series(py, s.clone(), lambda)
-            .map_err(|s| ComputeError(format!("{}", s).into()))?;
-        object.to_series(py, &POLARS, s.name())
-    })
-}
-
-fn python_function_caller_df(df: DataFrame, lambda: &PyObject) -> PolarsResult<DataFrame> {
-    Python::with_gil(|py| {
-        // create a PyDataFrame struct/object for Python
-        let pydf = PyDataFrame::new(df);
-        // Wrap this PyDataFrame object in the python side DataFrame wrapper
-        let python_df_wrapper = POLARS
-            .getattr(py, "wrap_df")
-            .unwrap()
-            .call1(py, (pydf,))
-            .unwrap();
-        // call the lambda and get a python side df wrapper
-        let result_df_wrapper = lambda.call1(py, (python_df_wrapper,)).map_err(|e| {
-            PolarsError::ComputeError(format!("User provided python function failed: {e}").into())
-        })?;
-        // unpack the wrapper in a PyDataFrame
-        let py_pydf = result_df_wrapper.getattr(py, "_df").map_err(|_| {
-            let pytype = result_df_wrapper.as_ref(py).get_type();
-            PolarsError::ComputeError(
-                format!("Expected 'LazyFrame.map' to return a 'DataFrame', got a '{pytype}'",)
-                    .into(),
-            )
-        })?;
-
-        // Downcast to Rust
-        let pydf = py_pydf.extract::<PyDataFrame>(py).unwrap();
-        // Finally get the actual DataFrame
-        let df = pydf.df;
-
-        Ok(df)
-    })
-}
-
-#[pyfunction]
-pub fn __register_startup_deps() {
-    if !registry::is_object_builder_registered() {
-        let object_builder = Box::new(|name: &str, capacity: usize| {
-            Box::new(ObjectChunkedBuilder::<ObjectValue>::new(name, capacity))
-                as Box<dyn AnonymousObjectBuilder>
-        });
-
-        let object_converter = Arc::new(|av: AnyValue| {
-            let object = Python::with_gil(|py| ObjectValue {
-                inner: Wrap(av).to_object(py),
-            });
-            Box::new(object) as Box<dyn Any>
-        });
-
-        registry::register_object_builder(object_builder, object_converter);
-        unsafe { python_udf::CALL_SERIES_UDF_PYTHON = Some(python_function_caller_series) }
-        unsafe { python_udf::CALL_DF_UDF_PYTHON = Some(python_function_caller_df) }
-    }
-}
